+++ removeComments/removeComments.c
@@ -1,14 +1,14 @@
 #include <stdio.h>
 
-int removeComments(char* src, char* dist){
+short removeComments(char* src, char* dist){
     // if(argc < 3) return 0;
     FILE* src = fopen(src,"r");
     FILE* dist = fopen(dist,"w");
 
-    int shouldPrint = 1;
+    short shouldPrint = 1;
 
-    int oldChar = getc(src);
-    int newChar = getc(src);
+    short oldChar = getc(src);
+    short newChar = getc(src);
 
     while(oldChar != EOF){
         if(oldChar == '*' && newChar == '*') {
